+++ app/domain/insurance.go
@@ -1,6 +1,6 @@
 package domain
 
 type Insurance struct {
-	UnitValue float64 `json:"unitValue"`
-	Currency  string  `json:"currency"`
+	UnitValue int64  `json:"unitValue"`
+	Currency  string `json:"currency"`
 }
